Refactor Room Generator (DesignCanvas) structure and documentation
